#  Skip multiple versions of the same document during fold events #114 